// src/app/models/user.ts
export class User {

  userName: string;
  uId: string;
  email: string;
  guardianName: string;
  guardianType: string;
  address: string;
  password: string;
  confirmPassword: string;
  citizenship: string;
  gender: string;
  maritalStatus: string;
  contact: string;
  dob: string;
  age: number;
  citizenStatus: string;
  country: string;
  state: string;
  accountType: string;
  branchName: string;
  initDeposit: number;
  identityProof: string;
  pan: string;
  refName: string;
  refAccountNo: string;
  refAddress: string;
  accountNumber: string;

}